+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Changelog
   - Go to quiz button for hosts to visit the page of quizzes in the queue.
   - Controls for hosts to toggle and customise interval for auto changing to
     next quiz.
+    - Countdown under leaderboard displaying time left until change to next
+    quiz.
+    - Button to cancel the countdown for hosts.
 
 [v2.2.0] - 2021-08-16
 ---------------------